requirements.txt
asgiref==3.7.2
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
inflection==0.5.1
isort==5.12.0
mccabe==0.7.0
packaging==23.1
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
